refactor: Passenger가 나이 정보를 가지고 있도록, AgeDiscountFarePolicy가 계산의 책임을 가지고 있도록 변경
