+++ src/main/java/subway/domain/fare/AgeDiscountFarePolicy.java
@@ -6,6 +6,8 @@ public class AgeDiscountFarePolicy implements FarePolicy {
 
     @Override
     public int calculate(final Path path, final Passenger passenger, final int fare) {
-        return passenger.calculateFare(fare);
+        final AgeGroup ageGroup = passenger.calulateAgeGroup();
+        final int discountedFare = fare - ageGroup.getDiscountFare();
+        return (int) (discountedFare - discountedFare * ageGroup.getDiscountRatio());
     }
 }
