+++ src/test/java/subway/domain/fare/AgeDiscountFarePolicyTest.java
@@ -15,10 +15,24 @@ class AgeDiscountFarePolicyTest {
     private FarePolicy farePolicy = new AgeDiscountFarePolicy();
 
     @Test
-    void 나이에_따라_금액을_계산한다() {
+    void 성인의_경우_할인이_적용되지_않는다() {
         // given
         final Path path = new Path(Collections.emptyList());
-        final Passenger passenger = new Passenger(17);
+        final Passenger passenger = new Passenger(19);
+        final int fare = 1250;
+
+        // when
+        final int result = farePolicy.calculate(path, passenger, fare);
+
+        // then
+        assertThat(result).isEqualTo(1250);
+    }
+
+    @Test
+    void 청소년의_경우_350원을_할인한_금액에서_20퍼센트가_할인된다() {
+        // given
+        final Path path = new Path(Collections.emptyList());
+        final Passenger passenger = new Passenger(18);
         final int fare = 1250;
 
         // when
@@ -26,5 +40,31 @@ class AgeDiscountFarePolicyTest {
 
         // then
         assertThat(result).isEqualTo(720);
+    }
+
+    @Test
+    void 어린이의_경우_350원을_할인한_금액에서_50퍼센트가_할인된다() {
+        final Path path = new Path(Collections.emptyList());
+        final Passenger passenger = new Passenger(6);
+        final int fare = 1250;
+
+        // when
+        final int result = farePolicy.calculate(path, passenger, fare);
+
+        // then
+        assertThat(result).isEqualTo(450);
+    }
+
+    @Test
+    void 유아의_경우_무료다() {
+        final Path path = new Path(Collections.emptyList());
+        final Passenger passenger = new Passenger(5);
+        final int fare = 1250;
+
+        // when
+        final int result = farePolicy.calculate(path, passenger, fare);
+
+        // then
+        assertThat(result).isZero();
     }
 }
